Move renderer to examples
We should only include generic renderers in the main source code.
